requirements.txt
python-dotenv
requests
titlecase
